Validation

+++ src/main/java/Service/ProductService.java
@@ -2,23 +2,31 @@ package Service;
 
 import DAO.ProductDAO;
 import DTO.ProductDTO;
-import DTO.ProductUpdateDTO;
 import Model.Product;
+import Validators.ProductValidator;
 
-import java.sql.Connection;
 import java.sql.SQLException;
+import java.time.Instant;
 import java.util.List;
+import java.util.Map;
 
 public class ProductService {
 
     private final ProductDAO productDAO = new ProductDAO();
+    private final ProductValidator validator = new ProductValidator();
 
-    public boolean addProduct(ProductDTO productDTO) throws SQLException {
+    public Map<String, String> validate(ProductDTO product) throws SQLException {
+        return validator.validate(product);
+    }
+
+    public Product addProduct(Product productDTO) throws SQLException {
         Product product = new Product.Builder()
                 .name(productDTO.getName())
                 .cost_price(productDTO.getCost_price())
                 .selling_price(productDTO.getSelling_price())
-                .opening_stock(productDTO.getOpening_stock())//stock in hand = opening stock during product creation
+                .opening_stock(productDTO.getOpening_stock())
+                .created_at(Instant.now().getEpochSecond())
+                .updated_at(Instant.now().getEpochSecond())
                 .build();
 
         return productDAO.insert(product);
@@ -28,16 +36,22 @@ public class ProductService {
         return productDAO.getAllRows();
     }
 
-    public Product getProductById(int id) throws SQLException {
+    public Product getProductById(Integer id) throws SQLException {
         return productDAO.findProduct(id);
     }
 
-    public boolean updateProduct(ProductUpdateDTO dto) throws SQLException {
+    public Product updateProduct(Product dto) throws SQLException {
         return productDAO.updateProduct(dto);
     }
 
-    public boolean deleteProduct(int id) throws SQLException {
+    public boolean deleteProduct(Integer id) throws SQLException {
+        if (productDAO.isProductUsed(id)) {
+            throw new IllegalStateException("Cannot delete: Product (PRO- " + id + ") is used in purchase or sales transactions.");
+        }
         return productDAO.deleteProduct(id);
     }
 
+    public boolean exists(int id) throws SQLException {
+        return productDAO.findProduct(id)!=null;
+    }
 }
